sdfgsfg

+++ lib/Clock_utils/Clock_utils.h
@@ -8,7 +8,7 @@ class Clock_utils
 public:
 	Clock_utils();
 	void init();
-	uint64_t get_t_now_s();
+	double get_t_now_s();
 	double get_t_now_ms();
 	double get_t_now_us();
 	void reset();
